readyup|refactor: Group kit checks and move helpers into test-utils
The modules the kits share now sit in `.readyup/kits/checks/` rather than `lib/`, naming what they hold instead of the compile rule that puts them there. That rule is now stated in `.config/readyup.config.ts` alongside the setting that creates it: every `.ts` directly under `kits/` is compiled as a kit, so anything shared has to sit one level down. Within `kit-layout.ts`, `ManifestEntry` sorts among its peer exports rather than under the paths heading.

Test helpers move from `__tests__/helpers/` to `test-utils/kits/`, outside the test tree so coverage can measure them, and take names describing their subject: `checklist-results.ts`, `project-fixture.ts`, and `loadOwnKit.ts`.

Coverage reaches `.readyup/kits/` and `test-utils/`, which the inherited `src`-only glob left unmeasured, and `tsconfig.json` includes both so the kit sources are typechecked directly rather than only through the tests importing them.
