+++ packages/readyup/.readyup/kits/checks/buildSelfContainmentChecks.ts
