+++ packages/readyup/__tests__/kits/default.test.ts
@@ -6,9 +6,9 @@ import process from 'node:process';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
 import type { RdyResult } from '../../src/types.ts';
-import { pickResult, runChecklist } from '../helpers/kitChecklistResults.ts';
-import { FIXTURE_KITS_DIR, writeKit, writeKitManifest, writeRdyConfig } from '../helpers/kitProjectFixture.ts';
-import { loadOwnKit } from '../helpers/loadOwnKit.ts';
+import { pickResult, runChecklist } from '../../test-utils/kits/checklist-results.ts';
+import { loadOwnKit } from '../../test-utils/kits/loadOwnKit.ts';
+import { FIXTURE_KITS_DIR, writeKit, writeKitManifest, writeRdyConfig } from '../../test-utils/kits/project-fixture.ts';
 
 /**
  * Covers the `default` kit readyup publishes, against fixture projects in a temp directory.
