+++ packages/readyup/tsconfig.json
@@ -1,5 +1,14 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
   "extends": "../../tsconfig.json",
-  "include": ["src/**/*.ts", "src/**/*.d.ts", "__tests__/", ".config/**/*.ts", "config/", "*.ts"],
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.d.ts",
+    "__tests__/",
+    "test-utils/",
+    ".config/**/*.ts",
+    ".readyup/kits/**/*.ts",
+    "config/",
+    "*.ts",
+  ],
 }
