+++ packages/readyup/vitest.config.ts
@@ -13,6 +13,12 @@ import { defineVitestConfig } from '@williamthorsen/nmr/vitest';
 // The aliases let a test import this package's own kits, which reach readyup by package name the way a consumer's kit
 // does, without a prior build standing between the test and the source. Longest specifier first: an alias on `readyup`
 // also matches `readyup/check-utils`, and the first match wins.
+//
+// The coverage entries extend the inherited `src`-only glob, which would otherwise leave both trees unmeasured. Kits
+// are authored under `.readyup/` rather than `src/` because that is the directory readyup itself reads, and helpers
+// under `test-utils/` are measured on purpose: an uncovered helper is why they sit outside `__tests__/` at all.
+// The text reporter prints no `test-utils` group even so; the files are in the coverage data, as the json and html
+// reporters show, so this is a reporter display gap rather than a glob that failed to match.
 export default defineVitestConfig({
   project: { env: { RDY_STYLE: 'rich' } },
   root: {
@@ -21,6 +27,11 @@ export default defineVitestConfig({
         { find: 'readyup/check-utils', replacement: path.resolve(import.meta.dirname, 'src/check-utils/index.ts') },
         { find: 'readyup', replacement: path.resolve(import.meta.dirname, 'src/index.ts') },
       ],
+    },
+    test: {
+      coverage: {
+        include: ['.readyup/kits/**/*.ts', 'test-utils/**/*.ts'],
+      },
     },
   },
 });
